reconciler/clusterworkspaces: rename to workspaces
