+++ pkg/reconciler/tenancy/workspace/workspace_reconcile_metadata_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package clusterworkspace
+package workspace
 
 import (
 	"context"
